+++ src/api/v3/role/types.rs
@@ -121,21 +121,24 @@ impl From<RoleListParameters> for types::RoleListParameters {
     }
 }
 
-///Role create request body
+/// Role create request body.
 #[derive(Clone, Debug, Default, Deserialize, PartialEq, Serialize, ToSchema, Validate)]
 pub struct RoleCreate {
-    /// Role domain ID
+    /// The role description.
     #[serde(skip_serializing_if = "Option::is_none")]
-    #[validate(length(max = 64))]
-    pub domain_id: Option<String>,
-    /// Role name
-    #[validate(length(max = 255))]
-    pub name: String,
-    /// Role description
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[validate(length(max = 255))]
+    #[validate(length(min = 1, max = 255))]
     pub description: Option<String>,
-    /// Extra attributes
+
+    /// The domain ID of the role.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[validate(length(min = 1, max = 64))]
+    pub domain_id: Option<String>,
+
+    /// The role name.
+    #[validate(length(min = 1, max = 255))]
+    pub name: String,
+
+    /// Extra attributes for the role.
     #[serde(flatten, skip_serializing_if = "Option::is_none")]
     pub extra: Option<Value>,
 }
@@ -143,11 +146,11 @@ pub struct RoleCreate {
 impl From<RoleCreate> for types::RoleCreate {
     fn from(value: RoleCreate) -> Self {
         Self {
-            id: Some(String::new()),
-            domain_id: value.domain_id,
-            name: value.name,
             description: value.description,
+            domain_id: value.domain_id,
             extra: value.extra,
+            id: None,
+            name: value.name,
         }
     }
 }
