+++ src/assignment/backend.rs
@@ -40,7 +40,7 @@ pub trait AssignmentBackend: DynClone + Send + Sync + std::fmt::Debug {
     /// Set config
     fn set_config(&mut self, config: Config);
 
-    /// Create Role
+    /// Create Role.
     async fn create_role(
         &self,
         state: &ServiceState,
